requirements.txt
pandas
requests
beautifulsoup4
pyyaml
lxml
openpyxl
streamlit
plotly